pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Maksim ilyukovich <m.ilyukovich@tinkoff.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
django-bootstrap4 = "^22.1"
django-filter = "^22.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"
gunicorn = "^20.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

